Add more typing to plots
This patch adds more typing to plots by turning Axis, Plots and Layouts into type constructors.

- Axis x now represents an axis for data of type x

- Plot x y represents a plot of data with independent variables of type x and dependent 
variables of type y.

- Layout1 x y y' represents a layout of multiple plots, all which must share the same 
independent variable type (x).  All plots must have one of two dependent variable types, either 
y, which will be shown on the Left axis, or y', which will be shown on the right Axis.

These types x, y and y' are intendend to be instances of my new PlotValue class.  Currently 
PlotValue has only one method:

 toValue :: a -> Double

I intend on expending this class to contain all the information needed for auto axis generation.

A new LogValue type is created containing a Double.  The autoScaledLogAxis generates axis for 
this type.  The instance of PlotValue for this type is

 toValue (LogValue x) = log x

This patch makes some minor changes to the interface.

plot_points_values and plot_lines_values now take pairs (x,y) instead of Points as values.  This 
actually makes the interface more uniform with plot_fillbetween_values and also make it more 
natural to interface with.

Secondly layout1_plots has a different type.  Instead of specifying

("am",HA_Bottom,VA_Left,(toPlot sinusoid1)),

one simply says

("am points", Left (toPlot sinusoid2)).

Similarly Right can be use to plot against the right axis.  It is now impossible to have 
different independent axis for the top and bottom axis.  However one can reuse much of the 
existing work to create a new layout that allows for different independent axis, but I don't 
think people really want that.  

It should be possible to have different labels for the top and bottom axis, so long as they 
underlying types of the axis are still the same.  I have no tests of this at the moment though.

Unfortuantely, I had to change the simple interface as well.  Because Layout1 is now a type 
constructor rather than a type, the old technique for multiple arguments does not work.  I 
created a newtype Layout1DDD for Layout1 Double Double Double.  To get a Layout1 from plot you 
need to unwrap Layout1DDD by passing it through plotLayout.  

Layout1DDD is an instance of ToRenderable, so if all you want to do is render a plot, you don't 
need to change anything.

I believe the simple interface can be modified to support ploting arbitrary types.  However, 
such a change is not a trival task, so I do not do it in this patch.