Split into two packages Chart & Chart-gtk
This removes the use of the use-gtk flag in the cabal file. The two
packages now live in subdirectories chart and chart-gtk. The simple
module needed to be split, as it contained GTK specific code. The
tests also needed to be split and refactored: The standard tests can
no longer be output to windows - they must be written to some sort of
file.
    
A trivial top level makefile was added to assist in building the two
packages and running tests.