BUGFIX: textDrawRect
The bounding-box calculation had the Y axis reversed, so the box was below the label. I have left in a commented-out version of drawTextA which adds in the bounding box, since there are differences between PNG/Cairo and SVG/Diagrams output, e.g. test8, which looks like a SVG rendering issue but I have not looked into it too deeply (you also see this difference in the misc1* tests).