Switch to scientific notation for large or small labels
First pass at fixing #121.

This will use scientific notation for all or none of the labels,
because I think it looks better that way. It also will apply an offset
too all the labels if the dynamic range is too small compared to the mean.

I includes doctests which are run with

    > runghc Doctests.hs

from the top-level git directory. If there is an unsatisfactory result from this
new function, a new doctest should be added.